+++ src/Layout.tsx
@@ -3,8 +3,11 @@ import Alert from 'src/components/common/Alert'
 import Footer from 'src/components/common/Footer'
 import Navbar from 'src/components/common/Navbar'
 import Toast from 'src/components/common/Toast'
+import RegionalizationBar from 'src/components/regionalization/RegionalizationBar'
+import RegionalizationModal from 'src/components/regionalization/RegionalizationModal'
 import { useUI } from 'src/sdk/ui'
 import type { PropsWithChildren } from 'react'
+import { useModal } from 'src/sdk/ui/modal/Provider'
 
 import 'src/styles/pages/layout.scss'
 
@@ -12,27 +15,38 @@ const CartSidebar = lazy(() => import('src/components/cart/CartSidebar'))
 
 function Layout({ children }: PropsWithChildren<unknown>) {
   const { displayMinicart } = useUI()
+  const { isRegionalizationModalOpen, setIsRegionalizationModalOpen } =
+    useModal()
 
   return (
-    <div id="layout">
-      <Alert>
-        Get 10% off today:&nbsp;<span>NEW10</span>
-      </Alert>
+    <>
+      <div id="layout">
+        <Alert>
+          Get 10% off today:&nbsp;<span>NEW10</span>
+        </Alert>
 
-      <Navbar />
+        <Navbar />
 
-      <main>{children}</main>
+        <main>
+          <RegionalizationBar classes="display-mobile" />
+          {children}
+        </main>
 
-      <Footer />
+        <Footer />
 
-      <Toast />
+        <Toast />
 
-      {displayMinicart && (
-        <Suspense fallback={null}>
-          <CartSidebar />
-        </Suspense>
-      )}
-    </div>
+        {displayMinicart && (
+          <Suspense fallback={null}>
+            <CartSidebar />
+          </Suspense>
+        )}
+      </div>
+      <RegionalizationModal
+        isOpen={isRegionalizationModalOpen}
+        onDismiss={() => setIsRegionalizationModalOpen(false)}
+      />
+    </>
   )
 }