+++ src/components/ui/Button/Button.tsx
@@ -3,10 +3,12 @@ import type { ReactNode } from 'react'
 import type { ButtonProps } from '@faststore/ui'
 
 export type Variant = 'primary' | 'secondary' | 'tertiary'
+export type Size = 'small' | 'regular'
 export type IconPosition = 'left' | 'right'
 
 export type UIButtonProps = {
   variant?: Variant
+  size?: Size
   inverse?: boolean
   icon?: ReactNode
   iconPosition?: IconPosition
@@ -16,6 +18,7 @@ type Props = ButtonProps & UIButtonProps
 
 function Button({
   variant,
+  size = 'regular',
   inverse,
   icon,
   iconPosition,
@@ -26,6 +29,7 @@ function Button({
     <UIButton
       data-fs-button
       data-fs-button-variant={variant}
+      data-fs-button-size={size}
       data-fs-button-inverse={inverse}
       {...props}
     >